Give control of fd provided in VHOST_USER_SET_BACKEND_REQ_FD
trait.VhostUserBackendMut.html#method.set_backend_req_fd provides a
wrapper (struct.Backend) around the fd in the VHOST_USER_SET_BACKEND_REQ_FD
message to handle the communication with the frontend.
struct.Backend implementation follows the vhost-user specs strictly, only
messages approved into the specs can be sent using this wrapper. However,
as a user of the vhost_user library I'd like to have the ability to send
custom messages, not yet in the specs, to the frontend. Passing the
ownership of the raw fd makes this possible without violating the specs.